+++ XochyTyanEngine.cpp
@@ -5,20 +5,20 @@
 int main(void)
 {
 
-    LinkedList<string> list;
+    LinkedList<string> *list = new LinkedList<string>();
     for (int i = 0; i < 10; i++)
     {
-        list.add("Info");
+        list->add("Info");
     }
-    list.toString(false);
-    list.remove("Info");
-    list.toString(false);
+    list->toString(false);
+    list->remove(0);
+    list->toString(false);
 
     WindowManager WINDOW_MANAGER;
     char name[] = "SomeWindow";
     WINDOW_MANAGER.init(name, 500, 500);
 
-    list.clear();
+    delete list;
 
     return 0;
 }
