+++ x64/Debug/XochyTyanEngine.log
@@ -1,2 +1,6 @@
-﻿  XochyTyanEngine.cpp
+﻿  LinkedList.cpp
+  XochyTyanEngine.cpp
+  Создание кода...
+  LINK : не найден или не выполнена сборка C:\Users\макс\YandexDisk\Engine\XochyTyanEngine\x64\Debug\XochyTyanEngine.exe при последней инкрементной компоновке; выполняется полная компоновка
+LINK : warning LNK4098: библиотека по умолчанию "MSVCRT" конфликтует с использованием других библиотек; используйте /NODEFAULTLIB:library
   XochyTyanEngine.vcxproj -> C:\Users\макс\YandexDisk\Engine\XochyTyanEngine\x64\Debug\XochyTyanEngine.exe
